INTERLOK-3260 NotReady shouldn't return a stack-trace
Change doResponse() so it takes a HTTP status as well.
Change the doResponse() so that sending a response doesn't cause
an exception since if it does, we're kinda screwed
If we're NotReady then this is just a normal doResponse with a 503
with simplistic JSON message - this will be "wrong" if someone has
a component with a unique-id with " in it.